Start all containers with the network
Always start the containers through dockStart.bat as it makes sure the network is created before the containers themselves because if you were to run the docker files before simultaneously, you'd run into the problem of creating multiple networks with the same name.
For the removal, it is currently a workaround. If you find a better way, please do so. First we remove the containers from the network and then they are decomposed with the network itself.

-- Maybe should've done the files for git